Create README.md
A web application that allows users to create rooms where other users can join and add the list of items that are to be bought. Its key functionality is to let the users of the room to check whether an item is bought or not, and if bought, it lets the users to view the id of the user who bought it. While the authentication of users is backed up by passport.js , the authentication for rooms is done using bcrypt. Technologies Used: HTML, CSS, Bootstrap, JavaScript, Node.js,Express.js, MongoDB. URL of the website: http://sharepod.herokuapp.com/